Renamed Footer component to Contact
