+++ src/App.vue
@@ -4,7 +4,7 @@
 	  <Home />
 	  <About />
 	  <!-- <Portfolio /> -->
-	  <Footer />
+	  <Contact />
   </div>
 </template>
 
@@ -13,7 +13,7 @@ import Header from './components/Header.vue'
 import Home from './components/Home.vue'
 import About from './components/About.vue'
 import Portfolio from './components/Portfolio.vue'
-import Footer from './components/Footer.vue'
+import Contact from './components/Contact.vue'
 
 // import Showcase from './components/Showcase.vue'
 
@@ -24,7 +24,7 @@ export default {
     Home,
     About,
 	Portfolio,
-    Footer
+    Contact
     // Showcase,
   }
 }
